:lipstick: fix coding style
