+++ server/src/CMakeLists.txt
@@ -9,6 +9,7 @@ target_sources(zappy_server_lib
         ${CMAKE_CURRENT_SOURCE_DIR}/hello_world.c
         ${CMAKE_CURRENT_SOURCE_DIR}/inventory.c
         ${CMAKE_CURRENT_SOURCE_DIR}/main.c
+        ${CMAKE_CURRENT_SOURCE_DIR}/map_bis.c
         ${CMAKE_CURRENT_SOURCE_DIR}/map.c
         ${CMAKE_CURRENT_SOURCE_DIR}/server_bis.c
         ${CMAKE_CURRENT_SOURCE_DIR}/server.c
