+++ server/src/server_bis.c
@@ -11,27 +11,6 @@
 #include <netinet/in.h>
 #include <string.h>
 
-static void write_to_client(client_t *client)
-{
-    char *message = buffer_get_next(client->buffer_answered);
-
-    for (; message != NULL; message =
-        buffer_get_next(client->buffer_answered)) {
-        send(client->socket, message, strlen(message), 0);
-        free(message);
-    }
-}
-
-static void write_to_clients(server_t *server)
-{
-    client_t *client;
-
-    for (size_t i = 0; i < array_get_size(server->clients); i++) {
-        client = (client_t *)array_get_at(server->clients, i);
-        write_to_client(client);
-    }
-}
-
 static int create_socket(server_t *server)
 {
     struct sockaddr_in *addr = malloc(sizeof(struct sockaddr_in));
@@ -44,7 +23,7 @@ static int create_socket(server_t *server)
     addr->sin_port = htons(server->port);
     addr->sin_addr.s_addr = INADDR_ANY;
     if (setsockopt(server->fd, SOL_SOCKET, SO_REUSEADDR, &opt,
-                   sizeof(opt)) == -1)
+        sizeof(opt)) == -1)
         return -1;
     if (bind(server->fd, (struct sockaddr *)addr, sizeof(*addr)) == -1 ||
             listen(server->fd, MAX_USERS) == -1)
@@ -60,7 +39,7 @@ void destroy(server_t *server)
 }
 
 server_t *create_server(unsigned short port, array_t *teams,
-                        size_t map_size[2], size_t nb_max_clients)
+    size_t map_size[2], size_t nb_max_clients)
 {
     server_t *server = malloc(sizeof(server_t));
 
@@ -77,34 +56,15 @@ server_t *create_server(unsigned short port, array_t *teams,
     return server;
 }
 
-void tick(server_t *server, __suseconds_t time_since_last_tick)
+void add_client(server_t *server, int fd)
 {
-    int nb_ticks = time_since_last_tick / server->single_tick_time;
+    client_t *client = init_client(fd, UNKNOWN);
 
-    if (nb_ticks <= 0)
-        return;
-    if (time_since_last_tick == -1)
-        nb_ticks = 0;
-    for (int i = 0; i < nb_ticks; i++) {
-        game_tick(server->game);
-    }
-    write_to_clients(server);
-    server->nb_ticks += nb_ticks;
-}
-
-void fill_fd_set(server_t *server, fd_set *readfds, fd_set *writefds)
-{
-    client_t *client;
-
-    FD_ZERO(readfds);
-    FD_ZERO(writefds);
-    FD_SET(server->fd, readfds);
-    server->max_fd = find_max_fd(server);
-    for (size_t i = 0; i < array_get_size(server->clients); i++) {
-        client = (client_t *)array_get_at(server->clients, i);
-        FD_SET(client->socket, readfds);
-        FD_SET(client->socket, writefds);
-    }
+    client->socket = fd;
+    buffer_write(client->buffer_answered, "WELCOME\n");
+    array_push_back(server->clients, client);
+    if (fd > server->max_fd)
+        server->max_fd = fd;
 }
 
 void handle_new_connections(server_t *server, fd_set *readfds)
@@ -116,5 +76,23 @@ void handle_new_connections(server_t *server, fd_set *readfds)
     if (FD_ISSET(server->fd, readfds)) {
         new_fd = accept(server->fd, (struct sockaddr *)&addr, &size);
         add_client(server, new_fd);
+    }
+}
+
+void remove_client(server_t *server, int fd)
+{
+    client_t *client;
+    size_t size = array_get_size(server->clients);
+
+    for (size_t i = 0; i < size; i++) {
+        client = (client_t *)array_get_at(server->clients, i);
+        if (client->socket == fd) {
+            array_remove(server->clients, i);
+            break;
+        }
+    }
+    if (fd == server->max_fd) {
+        server->max_fd = 0;
+        server->max_fd = find_max_fd(server);
     }
 }
